ajustes para el pago

+++ app/init_db.py
@@ -1,12 +1,12 @@
-from database.db import Base, engine
-from models.trips.trip_db import Trip
-from models.travel_tickets.travel_ticket_db import Ticket
-from models.users.user_db import User
+from app.database.db import Base, engine
+from app.models.travel_tickets.travel_ticket_db import Ticket
+from app.models.trips.trip_db import Trip
 
-# Primero borramos todas las tablas
-Base.metadata.drop_all(bind=engine)
+def init_db(drop_first: bool = False):
+    if drop_first:
+        Base.metadata.drop_all(bind=engine)
+    Base.metadata.create_all(bind=engine)
+    print("DB inicializada.")
 
-# Luego las creamos nuevamente
-Base.metadata.create_all(bind=engine)
-
-print("Tablas borradas y creadas correctamente.")
+if __name__ == "__main__":
+    init_db(drop_first=True)
